backend/requirements.txt
scikit-learn
pandas
matplotlib
numpy
fastapi
uvicorn
joblib
sqlalchemy
pydantic
xgboost
PyMySQL
python-dotenv
